# docs/openapi_contract.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "YouTube Content Automation ERP API",
    "description": "Production-grade API-First Control Plane for YouTube Automation & Durable Workflows",
    "version": "1.0.0"
  },
  "paths": {
...
    "/youtube/{idea_id}/status": {
      "get": {
        "tags": [
          "YouTube & Publishing"
        ],
        "summary": "Get Youtube Status",
        "operationId": "get_youtube_status_youtube__idea_id__status_get",
        "parameters": [
          {
            "name": "idea_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Idea Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/YouTubeStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/youtube/{idea_id}/payload": {
      "get": {
        "tags": [
          "YouTube & Publishing"
        ],
        "summary": "Get Upload Payload",
        "operationId": "get_upload_payload_youtube__idea_id__payload_get",
        "parameters": [
          {
            "name": "idea_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Idea Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/YouTubeUploadPayloadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/youtube/{idea_id}/upload": {
      "post": {
        "tags": [
          "YouTube & Publishing"
        ],
        "summary": "Upload Video",
        "operationId": "upload_video_youtube__idea_id__upload_post",
        "parameters": [
          {
            "name": "idea_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Idea Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/YouTubeUploadRequest",
                "default": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/YouTubeUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "YouTubeStatusResponse": {
        "properties": {
          "idea_id": {
            "type": "integer",
            "title": "Idea Id"
          },
          "is_ready": {
            "type": "boolean",
            "title": "Is Ready"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "has_video": {
            "type": "boolean",
            "title": "Has Video"
          }
        },
        "type": "object",
        "required": [
          "idea_id",
          "is_ready",
          "has_video"
        ],
        "title": "YouTubeStatusResponse"
      },
      "YouTubeUploadPayloadResponse": {
        "properties": {
          "idea_id": {
            "type": "integer",
            "title": "Idea Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "tags": {
            "type": "string",
            "title": "Tags"
          },
          "category": {
            "type": "string",
            "title": "Category"
          }
        },
        "type": "object",
        "required": [
          "idea_id",
          "title",
          "description",
          "tags",
          "category"
        ],
        "title": "YouTubeUploadPayloadResponse"
      },
      "YouTubeUploadRequest": {
        "properties": {
          "video_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Path"
          }
        },
        "type": "object",
        "title": "YouTubeUploadRequest"
      },
      "YouTubeUploadResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "video_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Id"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "YouTubeUploadResponse"
      }
    }
  }
}